Limit total size of queued messages for p2p nodes (currently 1MB). Add exception types for bts::net. Re-enable and fix compile errors in bitshares_client_tests. Quiet a few compiler warnings

+++ tests/dev_fixture.hpp
@@ -168,7 +168,7 @@ struct chain_fixture
          
          config.names.push_back(delegate_account);
          config.balances.push_back( std::make_pair( pts_address(fc::ecc::public_key_data(delegate_account.owner)), 
-                                                    BTS_BLOCKCHAIN_INITIAL_SHARES/BTS_BLOCKCHAIN_NUM_DELEGATES) );
+                                                    (double)BTS_BLOCKCHAIN_INITIAL_SHARES/BTS_BLOCKCHAIN_NUM_DELEGATES) );
       }
 
       fc::json::save_to_file( config, clienta_dir.path() / "genesis.json" );
